// jobup-frontend/src/app/generated-sources/openapi/model/notificationDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface NotificationDto { 
    id?: string;
    recipientId?: string;
    recipientName?: string;
    senderId?: string;
    senderName?: string;
    refId?: string;
    refTitle?: string;
    type?: NotificationDto.TypeEnum;
    message?: string;
    actionUrl?: string;
    read?: boolean;
    createdAt?: string;
}
export namespace NotificationDto {
    export type TypeEnum = 'POST_LIKED' | 'POST_COMMENTED' | 'POST_SAVED' | 'PROPOSAL_RECEIVED' | 'PROPOSAL_ACCEPTED' | 'PROPOSAL_DECLINED' | 'DEAL_CONFIRMED' | 'DEAL_IN_PROGRESS' | 'DEAL_COMPLETED' | 'DEAL_CANCELLED' | 'RATING_ADDED';
    export const TypeEnum = {
        PostLiked: 'POST_LIKED' as TypeEnum,
        PostCommented: 'POST_COMMENTED' as TypeEnum,
        PostSaved: 'POST_SAVED' as TypeEnum,
        ProposalReceived: 'PROPOSAL_RECEIVED' as TypeEnum,
        ProposalAccepted: 'PROPOSAL_ACCEPTED' as TypeEnum,
        ProposalDeclined: 'PROPOSAL_DECLINED' as TypeEnum,
        DealConfirmed: 'DEAL_CONFIRMED' as TypeEnum,
        DealInProgress: 'DEAL_IN_PROGRESS' as TypeEnum,
        DealCompleted: 'DEAL_COMPLETED' as TypeEnum,
        DealCancelled: 'DEAL_CANCELLED' as TypeEnum,
        RatingAdded: 'RATING_ADDED' as TypeEnum
    };
}